Only delete cache on Github Action

+++ src/__main__.py
@@ -1,3 +1,4 @@
+import os
 import argparse
 from src.domains import DomainConverter
 from src.cloudflare import (
@@ -98,6 +99,9 @@ class CloudflareManager:
             self.cache["rules"] = []
             utils.save_cache(self.cache)
 
+def is_running_in_github_actions():
+    return os.getenv('GITHUB_ACTIONS') == 'true'
+
 def main():
     parser = argparse.ArgumentParser(description="Cloudflare Manager Script")
     parser.add_argument("action", choices=["run", "leave"], help="Choose action: run or leave")
@@ -106,7 +110,8 @@ def main():
     
     if args.action == "run":
         cloudflare_manager.update_resources()
-        utils.delete_cache()
+        if is_running_in_github_actions():
+            utils.delete_cache()
     elif args.action == "leave":
         cloudflare_manager.delete_resources()
     else:
